- Applied theme to remote branch - Fixed all references for the theme      >content_main      >content_db_list      >AndroidManifest      >...few other places?
The project does build, install, and run.
The back button used in the Db causes it to crash. Problem with the whenBackPressed() method

Branches probably are not something that I need, or is required, at this point.
Try to get rid of the 'remote' branch, Merge back into the master?
:: Remote AA595